force tests to run sequentially to stop issues with Current.Factory
