+++ tests/UmbMapper.Umbraco8.Tests/Properties/AssemblyInfo.cs
@@ -18,3 +18,4 @@ using System.Runtime.InteropServices;
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("1.0.0.0")]
 [assembly: AssemblyFileVersion("1.0.0.0")]
+[assembly: Xunit.CollectionBehavior(MaxParallelThreads = 1)]
